Make uniform a signed int again

+++ components/sceneutil/shadow.cpp
@@ -104,7 +104,7 @@ namespace SceneUtil
             stateset.setTextureAttribute(i, fakeShadowMapTexture,
                 osg::StateAttribute::ON | osg::StateAttribute::OVERRIDE | osg::StateAttribute::PROTECTED);
             stateset.addUniform(new osg::Uniform(
-                ("shadowTexture" + std::to_string(i - mShadowSettings->getBaseShadowTextureUnit())).c_str(), i));
+                ("shadowTexture" + std::to_string(i - mShadowSettings->getBaseShadowTextureUnit())).c_str(), static_cast<int>(i)));
         }
     }
